requirements.txt
jinja2
PyYAML
python-slugify
